+++ src/lib/settings/types/UserSettings.ts
@@ -2,8 +2,8 @@
 import { HexColorString } from 'discord.js';
 import { O } from 'ts-toolbelt';
 
-import { Quests } from '../../../commands/Minion/quest';
 import { BitField } from '../../constants';
+import { Quests } from '../../data/QuestExports';
 import { GearSetup } from '../../gear';
 import { CombatOptionsEnum } from '../../minions/data/combatConstants';
 import { PatchTypes } from '../../minions/farming';
