+++ src/tasks/minions/questingActivity.ts
@@ -1,74 +1,83 @@
-import { KlasaMessage, Task } from 'klasa';
+import { KlasaMessage, KlasaUser, Task } from 'klasa';
+import { Bank } from 'oldschooljs';
+import { resolveBank } from 'oldschooljs/dist/util';
 
-import { Emoji, MAX_QP } from '../../lib/constants';
+import { QuestList } from '../../lib/data/QuestExports';
 import { UserSettings } from '../../lib/settings/types/UserSettings';
 import { SkillsEnum } from '../../lib/skilling/types';
 import { QuestingActivityTaskOptions } from '../../lib/types/minions';
-import { rand, roll } from '../../lib/util';
+import { formatSkillRequirements } from '../../lib/util';
 import { handleTripFinish } from '../../lib/util/handleTripFinish';
+
+export async function completeUserQuestID(user: KlasaUser, questID: number) {
+	const mainQuest = QuestList.find(q => q.id === questID);
+	if (mainQuest) await user.addQP(mainQuest.rewards.qp);
+	return user.settings.update(UserSettings.Quests, questID, {
+		arrayAction: 'add'
+	});
+}
 
 export default class extends Task {
 	async run(data: QuestingActivityTaskOptions) {
-		const { userID, channelID } = data;
+		const { userID, channelID, questID } = data;
 		const user = await this.client.users.fetch(userID);
-		const currentQP = user.settings.get(UserSettings.QP);
-
-		// This assumes you do quests in order of scaling difficulty, ~115 hours for max qp
-		let qpRecieved = rand(1, 3);
-		if (currentQP >= 200) {
-			qpRecieved = 1;
-		} else if (currentQP >= 100) {
-			qpRecieved = rand(1, 2);
+		const quest = QuestList.find(q => q.id === questID)!;
+		// xp
+		if (quest.rewards?.xp) {
+			for (const [skill, amount] of Object.entries(quest.rewards.xp)) {
+				await user.addXP({ skillName: skill as SkillsEnum, amount });
+			}
 		}
-
-		const newQP = currentQP + qpRecieved;
-
-		// The minion could be at (MAX_QP - 1) QP, but gain 4 QP here, so we'll trim that down from 4 to 1.
-		if (newQP > MAX_QP) {
-			qpRecieved -= newQP - MAX_QP;
+		// items
+		if (quest.rewards?.items) {
+			await user.addItemsToBank(resolveBank(quest.rewards.items));
 		}
-
-		let str = `${user}, ${
-			user.minionName
-		} finished questing, you received ${qpRecieved.toLocaleString()} QP. Your current QP is ${
-			currentQP + qpRecieved
-		}.`;
-
-		const hasMaxQP = newQP >= MAX_QP;
-		if (hasMaxQP) {
-			str += `\n\nYou have achieved the maximum amount of ${MAX_QP} Quest Points!`;
+		// Custom reward
+		let customRewardString = '';
+		let rewardToCollect = '';
+		if (quest.rewards.customLogic) {
+			for (const reward of quest.rewards.customLogic) {
+				if (reward.type === 'direct_reward') {
+					const customReward = reward.function();
+					switch (customReward[0]) {
+						case 'xp': {
+							for (const [skill, amount] of Object.entries(customReward[1])) {
+								await user.addXP({ skillName: skill as SkillsEnum, amount });
+							}
+							customRewardString += formatSkillRequirements(customReward[1]);
+							break;
+						}
+						case 'item': {
+							const { itemsAdded } = await user.addItemsToBank(resolveBank(customReward[1]));
+							customRewardString += new Bank(itemsAdded).toString();
+							break;
+						}
+					}
+				} else if (reward.type === 'collect_reward') {
+					rewardToCollect += `You can collect another reward doing \`?q collect ${quest.name.toLowerCase()} ${
+						reward.id
+					}\`\n`;
+				}
+			}
 		}
-
-		await user.addQP(qpRecieved);
-		const herbLevel = user.skillLevel(SkillsEnum.Herblore);
-		if (herbLevel === 1 && newQP > 5 && roll(2)) {
-			await user.addXP({ skillName: SkillsEnum.Herblore, amount: 250 });
-			str += `${Emoji.Herblore} You received 250 Herblore XP for completing Druidic Ritual.`;
-		}
-
-		const magicXP = user.settings.get(UserSettings.Skills.Magic);
-		if (magicXP === 0 && roll(2)) {
-			await user.addXP({ skillName: SkillsEnum.Magic, amount: 325 });
-			str += `${Emoji.Magic} You received 325 Magic XP for completing Witch's Potion.`;
-		} else if (magicXP < 1000 && newQP > 15 && roll(2)) {
-			await user.addXP({ skillName: SkillsEnum.Magic, amount: 1000 });
-			str += `${Emoji.Magic} You received 1000 Magic XP for completing Fairytale I - Growing Pains.`;
-		} else if (user.skillLevel(SkillsEnum.Cooking) >= 40 && newQP > 50 && magicXP < 2500 && roll(2)) {
-			await user.addXP({ skillName: SkillsEnum.Magic, amount: 2500 });
-			str += `${Emoji.Magic} You received 2500 Magic XP for completing Recipe For Disaster (Lumbridge guide subquest).`;
-		}
+		await completeUserQuestID(user, quest.id);
+		const returnStr = `${user}, ${user.minionName} returned from its adventure, completing **${
+			quest.name
+		}**. You received the following rewards:\n${quest.rewards?.qp} QP (You now have ${user.settings.get(
+			UserSettings.QP
+		)} QP)${quest.rewards?.xp ? `\n${formatSkillRequirements(quest.rewards?.xp)}` : ''}${
+			quest.rewards?.items ? `\n${new Bank(quest.rewards?.items)}` : ''
+		}${customRewardString ? `\n${customRewardString}` : ''}${rewardToCollect ? `\n${rewardToCollect}` : ''}`;
 
 		handleTripFinish(
 			this.client,
 			user,
 			channelID,
-			str,
-			hasMaxQP
-				? undefined
-				: res => {
-						user.log('continued trip of Questing.');
-						return this.client.commands.get('quest')!.run(res as KlasaMessage, []);
-				  },
+			returnStr,
+			res => {
+				user.log('continued trip of Questing.');
+				return this.client.commands.get('quest')!.run(res as KlasaMessage, []);
+			},
 			undefined,
 			data,
 			null
